Rename files to snake-case.
