+++ bin/test-runner/tests/codec/work-result.ts
@@ -3,7 +3,7 @@ import fs from "node:fs";
 import type { HASH_SIZE } from "@typeberry/block";
 import type { Gas, ServiceId } from "@typeberry/block";
 import { CodecContext } from "@typeberry/block/context";
-import { WorkExecResult, WorkExecResultKind, WorkResult } from "@typeberry/block/work_result";
+import { WorkExecResult, WorkExecResultKind, WorkResult } from "@typeberry/block/work-result";
 import { type Bytes, BytesBlob } from "@typeberry/bytes";
 import { Decoder, Encoder } from "@typeberry/codec";
 import { json } from "@typeberry/json-parser";
