+++ packages/block/gaurantees.ts
@@ -2,7 +2,7 @@ import { type CodecRecord, codec } from "@typeberry/codec";
 import type { KnownSizeArray } from "@typeberry/collections";
 import type { TimeSlot, ValidatorIndex } from "./common";
 import { ED25519_SIGNATURE_BYTES, type Ed25519Signature } from "./crypto";
-import { WorkReport } from "./work_report";
+import { WorkReport } from "./work-report";
 
 export class ValidatorSignature {
   static Codec = codec.Class(ValidatorSignature, {
